Nuitka 0.3.9 pre3
Use one cached frame object/code object for each function to add tracebacks
when exceptions are raised. This largely speeds up the case where exceptions
are raised, now only a traceback object needs to be created.

Also avoid exporting the exception to Python and changing the traceback there,
do it on the C++ exception's traceback object instead. And inline the code to
add the traceback so it doesn't require the frame and code object to have the
line number correct. We can set it directly this way.