Added more tests:
Order in which dictionary keys and values are evaluated. Was correct, but shall
remain correct in upcoming changes.

Use of dict builtin to define a constant dictionary. Was not creating optimal
for code it yet.

Reference count test for assignments to multiple targets, with and without
unpacking. Covered by CPython test suite, but this should be more clearly
tested.

Comparison chains in all possible contexts. Before solving the TODO to no longer
use C++ static globals for them, cover all the cases, so nothing escapes. These
are rarely used and I am not sure the CPython test suite would cover them.