Nuitka 0.3.10 pre3
Enhanced builtin optimizations handling to be also able to deal with
dictionary star args. It requires the wraping of the node creation,
because dict takes a single unnamed positional, but optional parameter.

Also some pylint work for parameter parsing to produce less lengthy
code and use the templates for calls introduced previously.