Nuitka 0.3.10 pre4
Refactored scons handling to separate module, added inline copy of Scons
with a patch applied that should make it not require pywin32 on the Windows
platform.

The inline copy is only used if no /usr/bin/scons is found. It also serves
to lower the bar of entry to Nuitka. Packagers may want to not include it,
but instead to depend on scons >= 2.0.1 versions.