Nuitka 0.3.10pre5
- Corrected the behaviour of sys.exc_info() values for the case of nested
exceptions. There is now use of "FrameExceptionKeeper" when a try/except is
used that can preserve existing exceptions and restore as necessary.

- Removed the documentation that said scons is required. It's now included
and therefore always there.

- Added status information for the optimizations plan of README.txt to make
it easier to discover.

- Optimize the case of calling a function with only star list parameter as
it is really very trivial to implement.

- Removed TODO as it's now clear that weakrefs must be supported for the
generator functions, CPython does it too. We did, but were not sure it
was the right thing.

- Refactored exception raising code, so it is easier to understand, set
the traceback indicator inside functions if needed.

- Removed useless globals variable header file. As every module has it
now for itself, no need to have that file and moved the remaining parts
to better places.

- Much improved XML dump output. Intended for regression testing the
optimization changes. Gives a tree with roles.

- Changes all nodes to children having, there were a few holdouts. Now
the order of children is fixed, and every child has a name and a role
so a pretty tree can be made.

- Added mayHaveSideEffects and mayRaiseException where the knowledge that
something might change things behind your back or raise an exception can
be represented.

- Do away with lambda bodies, use function bodies instead. This allows to
do away with special code, and allows optimizations to work more generally
on functions and lambdas the same.

- Added infrastructure for simulation of operations, used for the constant
propagation of them.

- Added support for referencing builtin exceptions and for making their
objects as nodes. This leads to exception creation being optimized not
using global name lookups, but faster creation of just the right object.

- Treat the `` operator just like any other operator, optimize it as well.

- Optimize comparison functions with constant operators as well.

- Use assign targets in exception handlers, make handler an object, with
proper structure, instead of having lists with difficult to recognize
correspondance.

- Cleanup of Contexts, removing lots of unused stuff.

- Make it more clear where a generator expression is meant in the context
of code generation for contractions, these are not generators.

- Temporary variables for comparison chains are no longer module global,
but local to the containing function body, class body, etc.

- Removed the need for assigning the parent of whole trees. Operations that
change the tree now ensure it's correct.

- Added optimizations of try/except with known exception raises, and the
try/finally with no exception possible.

- Made the source executable with Python3.2 again, was broken. Doesn't work
with it at all though.

- Extended to coverage of optimizing builtin exceptions in many cases, more
variants supported.