Detect when running on ARM and use our own getcontext/makecontext/swapcontext in that case.
The ARM libc appears to not implement the obsolete interface, so this was
taken from libtask by Russ Cox.

The code didn't work as was with current ARM, because the definition of
the register storage area changed, but the assembler code assumes other,
so makecontext looks really strange. The best about it is that it works.