Portability work for ARM.
The order in which arguments are evaluated is not specified for C++,
but is quite fixed for a given ABI. For x86/x64 is is right to left
whereas for ARM it is left to right.

This has an impact on the order in which Python expressions are
evaluated. To be compatible with CPython, we need to wrap argument
lists and calls into macros that reverse them as necessary.

The change has the benefit that the code can now express that order
instead of always using the reversed order directly.

Currently MAKE_LIST, MAKE_TUPLE and MAKE_DICT have been addressed,
which covers a lot already. Function calls have their arguments in
the correct order with this, but e.g. CALL_FUNCTION variants need
a similar change to evalute the called expression, keyword args and
plain args, star list/dict arg in the correct order.

As a positive side-effect, a double reversal done in the case of
decorators has become unnecessary and removed.