More ARM portability work and a solved TODO about source code layout related to the reversal macros used.
All CALL_FUNCTION variants now evalutate in the correct order as well,
achieved with a macro definition that potentially reverses the arguments
of the caller as necessary.

This macro trick avoids the pollution of the code with manual reverse
macro usages.

As to reverse macros, now those for 2 to 6 arguments are created in any
case, there is no point to trace why they exist. Much code is going to
need them anyway.

And these macros are now generated into in a file of its own right, so
that is now cleaner, and code can use just these.