Special case all variants of function calls. Consistency cleanups.
Only a few combinations of calls had special code, and the typical
short cases of no args or only positional or only keyword args was
not generating readable code.

To enhance this, and to generally review the function call code, I
completed this, and did a few consistency improvements and potential
order bug fixes on the way.