Good bye org-mode, welcome rst.
We never took advantage of org-mode, whereas with rst it's really simple
to produce nice looking PDF and still have a readable ASCII format. For
me org-mode turned out too hard to learn.

Also restructured the document so that important sections come earlier,
and some content fixups, spelling, etc.

The rst is a child of Python and easier to read too.