Fix, the "__import__()" wasn't recursed into and could cause a crash.
The issue was many fold:

1. Module recursion used outdated code that was using attributes of the
"builtin import" node, that no longer existed, after previous refactorings.
2. The module recursion was not run again when a "builtin import" node was
created.
3. Not embedded the module was not detected by the tests.
4. The "builin import" node is no longer needed at all. There is now a
"import module expression" node that does its trick, and actually more.

This commit removed the "builtin" node and the obsolete code, points 1
and 4. Other changes in the development tree have made sure the issue
is detected by addressing 2 and 3.

Thanks go to Christopher Toth for reporting the issue and helping to
debug it.