Fix, assigning to True, False and None didn't work.
Solving the TODO to move this to optimization, we can check if it is
allowed to make this optimization, or if these names were assigned to,
so it's now safe.

Was causing a crash, when assigning to "True", "False" or "None". For
the boolean names, some very old code actually does these things.