More Python3 portability work, it even runs a few programs now.
The solution of Issue#8 (Python3 support) will take some more time, but
it appears, we will be getting there.

The implementation of PRINT_ITEM_TO is using the print builtin and ought
to be used on Python2 as well, but uncovered a TODO, namely that creating
dictionaries with MAKE_DICT is not yet portable enough. Once that is the
case, we could remove the non-builtin using code, as it's available on
CPython2.6 as well.

The basic tests are passed through 2to3 when Python3 is used.