Fix, detect collisions of upper and lower case module names.
While these collisions are not portable, they may still exist, and
they can be be done portable, when you have a module and a package
of the same name.

The real world example of it was "cython" which uses "Cython".

The resolution is to go over all modules (packages are modules too)
and collect the colliding filenames with "os.path.normcase()" and
if the filenames collide, then to enumerate them in their filenames.