According to Python Issue#13735, we need to call "pickletools.optimize()" for a stable result.
The CPython2.7 appears to make an optimization that can only be done if
no other is holding a reference it seems. With the above call it is done
for all versions.

It seems that this is a good idea anyway, as it makes the stream the most
efficient, so enable this by default.