Use "builtin name reference" node for "staticmethod" and constant "type" lookup results.
This solves TODOs about not very safe usages of "staticmethod" automatically added
for "__new__" and "type( 'lala' )" to "str" builtin values.

It will be instrumental for working on better handling of builtin calls later on as
well.

What's probably missing now, is the optimization for builtin types, to be looked up
faster and impossible to override.

The approach will be to have first builtin name references from read only module
variable names, which then are converted to calls in a later step. Currently it's
in one go, but that only complicates the optimization and doesn't play as nice as
it could.