Workaround for Issue#26, added a "two_step_execution" mode, where "--execute" is not used.
This ought to allow to run the tests unaffected by this MinGW bug, which has also been
raised, as it's a compiler bug (in my eyes) that exec doesn't forward the real exec
result. But it may also be a limitation of the Windows platform, so a workaround was
needed for now.