Refactor comparison node to always have only 2 operands, and solve comparison chains by introducing temp variables and assignment expressions.
This reduces the burden on code generation to provide "expression_temps" which could only
be of "PyObject *" type ever (in principle now they could change type easily), and had to
be provided more manually.

It reduces the burden on optimizations, because now the short circuit evualation nature
of chains is expressed by using "and" operation, which does this.

With value propagation in place, these temp variables may be eliminated later on entirely,
making the structure totally normal then.

So this removes a major wart of the node tree again and should make comparisons easy to
compute at compile time even if they are chains.