Remove useless "new_variable" tag, it's no longer used.
In the mean time, builtin references nodes are created where previously
new variable references had been used, which then needed to trigger the
resolution of these. Since that is no longer allowed, the TODO can be
solved, and this can can be removed.

Also check that tags checked for actually exist, that may catch errors
in checking code.