Make sure the sconsign path exists and is absolute, or else a Scons trace is triggered abouts its creation.
For some reason, the scons contains this code:

    def SConsignFile(self, name=".sconsign", dbm_module=None):
        if name is not None:
            name = self.subst(name)
            if not os.path.isabs(name):
                name = os.path.join(str(self.fs.SConstruct_dir), name)
        if name:
            name = os.path.normpath(name)
            sconsign_dir = os.path.dirname(name)
            if sconsign_dir and not os.path.exists(sconsign_dir):
                self.Execute(SCons.Defaults.Mkdir(sconsign_dir))
        SCons.SConsign.File(name, dbm_module)

This commit is to avoid to trigger the conditions under which the creation of
the directory is performed, because it gives a trace output on stderr, which
breaks the test process.

Calling "SCons.SConsign.File" directly would perhaps be even better, but the
access to it is well hidden.