Minor Optimization: Avoid using "PyModule_GetDict" to access module dictionary.
Do it directly, as we do in other places already, slightly faster and
more friendly to the optimizer (no external call) for "globals" code
and users of it as "import" statements.

Stumbled over it in valgrind analysis only for presence, not for actual
cycles. The effect of course is barely there.