Add command line flag --user-plugin
Right now, all the plugins have to be in the source code of Nuitka.
Adding the flag would allow Nuitka users to write their own plugins.
And the source code of the plugin can be put anywhere.
It's even possible to write per project user plugin.

The flag can be used multiple times.