Exclude common glibc libraries from standalone build
All libraries which come with glibc must be in sync to prevent
compatibility issues when distributing the executables to other
machines.

This means we should either not exclude any glibc libs and ship them all
(which may cause even more problems) or just exclude all of them
since they are related to and highly dependent  each other.