[TEST] RSA PyPI standalone test for Nuitka
This is done by testing the encryption and decryption
components of the rsa package.

Basically, it involves the creation of a public/private
key pair which is used to encrypt a given message using
the public key, sends it to the recipient and uses the
private key to decrypt the encrypted message at the
receiving end.

An assertion is used to check that the message stays the
same after the encryption-decryption process.

This contributes to https://github.com/Nuitka/Nuitka/issues/229.
Algorithm
=========

- Create Private/Public key pair
- Encode with UTF-8 and encrypt the message then send
- Decrypt the message and decode it using UTF-8
- Assert to make sure the message at the receiving end
  matches what was sent.

<create (priv/pub) key> ---> <encode and encrypt message>
---> <decrypt and decode message> ---> <assert message>

Signed-off-by: tab1tha <tambetachere@gmail.com>