[TEST]Nuitka standalone test for Pyparsing PyPI package.
The pyparsing module provides a library of classes that
client code uses to construct the grammar directly in python code.

The Pyparsing package is tested by using it to parse a string
of the form <salutation>,<greeting>! ,and then checking that
the result type is as expected( In this case,result type is a list.

Contributes to https://github.com/Nuitka/Nuitka/issues/229
Signed-off-by: tab1tha <tambetachere@gmail.com>