[Enhancement] CLI argument handling for standalone binaries.
This commit implements argument parsing in the standalone test
runner 'run_all.py'.

A demonstration of this implementation is done in the attached
standalone test for Nuitka compatibility with the click PyPI
package.

Command line argument parsing is implemented by using the argparse
module(because optparse is no longer being maintained).

===========================Algorithm=============================
The arguments that the program is willing to accept are defined in the
program code using the add_argument() method of the argparse module.

The arguments are passed when the program is invoked. However, these
arguments are interpreted(parsed) by the runner using the parse_args()
method of argparse.
parse_args() converts strings to objects and assigns them as attributes
of the namespace. It returns the populated namespace.
These attributes can then be picked up by the program so that the necessary
actions can be carried out(for example; performing a calculation or printing
a string)
=================================================================

The parse_args() method receives two parameters: args, and namespace.
args contains the list of strings to be parsed and its default is gotten from
sys.argv while namespace is an object to store the attributes produced; its
default is a new empty Namespace object.
Signed-off-by: tab1tha <tambetachere@gmail.com>