add support for singed onefile-application on windows
Parse the pe header using MapAndLoad() system function
to find out whether the application has a trailling certificate
rather then assume that the start position of the container is at the
very end of the file.
Convert the unicode path to its short form before transforming it
to ansi as suggested by kayhayen.

Close: #1154
Signed-off-by: Martin Gysel <me@bearsh.org>