Standalone: Resolve executable called through symbolic link (#1722)
For all platforms (tested on Linux, Windows and macOS), adds a mechanism to
resolve any symbolic links in the path of the executable when determining the
directory of the binary. This allows for a standalone dist directory to be
deployed to a system and a symbolic link created to the executable from a
directory already on the system PATH to that dist directory.

Without this, resolution of shared objects and Python package resources will
fail since it sets the Python directory to that of the link, not the file the
link targets.

For all POSIX like systems, this makes use of realpath() to fully resolve the
path. On Windows similar is achieved with GetFinalPathNameByHandle() (with a W
suffix for wchar support.)