do not check known kernel shared libraries for relative paths
Nuitka contains a list of known "kernel" shared libraries that are not included.
Since these are not included, we do not need to ensure that they don't contain
relative paths.

Without this patch I got this error:
FATAL: Error: Found a dependency with a relative path. Was a dependency copied to dist early? linux-vdso.so.1