Onefile: Preserve the original argv[0] in __compiled__.onefile_argv0
In commit 9b0406b2a ("Fix, need to make sure sys.argv[0] is absolute
for best usability"), the original argv[0] as received by the onefile
binary is thrown away, breaking applications needing this information.

To support these cases, record the original argv[0] in the
NUITKA_ONEFILE_ARGV0 environment variable, for exposure via
__compiled__.onefile_argv0. The field is made None if the onefile mode
is not in use, or if the binary is not launched via the onefile
bootstrapper.