Allow disabling nuitka in pyproject.toml builds.
When using `pyproject.toml` it is sometimes helpful to be able to
switch between `setuptools`-based build and a Nuitka build without
modifying the `pyproject.toml`.

Here, a configuration setting `no-nuitka` is added, which can be set
to `true`, resulting in a bypass of nuitka during build.

This mechanism also works for the `build` package, using the
`--config-setting` command line option.

Fixes #1600.